+++ src/postup_commands.cpp
@@ -3,25 +3,23 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #include "postup_commands.hpp"
+#include <boost/process/v2/process.hpp>
+#include <boost/process/v2/stdio.hpp>
 
 void PostupCommands::add_postup_command(const std::string& command)
 {
     this->_postup_commands.push_back(command);
 }
 
-void PostupCommands::execute_commands() const
+void PostupCommands::execute_commands(const boost::asio::any_io_executor ex) const
 {
     for (const auto& postup_command : this->_postup_commands)
     {
         BOOST_LOG_TRIVIAL(debug)
             << "Execute post up command: " << postup_command << std::endl;
-        std::error_code ec;
-        boost::process::child child(
-            postup_command,
-            boost::process::std_out > boost::process::null,
-            boost::process::std_err > boost::process::null,
-            ec);
-        child.wait();
+        boost::process::process child(ex, postup_command, {}, boost::process::process_stdio{.in=nullptr, .out=nullptr, .err=nullptr});
+        boost::system::error_code ec;
+        child.wait(ec);
         BOOST_LOG_TRIVIAL(debug)
             << "Post up command result: " << child.exit_code() << std::endl;
         if (child.exit_code() != 0 || ec)