+++ src/postup_commands.hpp
@@ -12,7 +12,7 @@ class PostupCommands
 {
     public:
         void add_postup_command(const std::string& command);
-        void execute_commands() const;
+        void execute_commands(const boost::asio::any_io_executor ex) const;
 
     private:
         std::vector<std::string> _postup_commands;